refactor(config): into adatpers and simplifies config/ folder
