+++ src/config/index.ts
@@ -1,18 +1,4 @@
-import { createConfig, http } from "wagmi";
-import { activeChains } from "./chains";
-import { injected } from "wagmi/connectors";
-import { beexoConnector } from "./beexoConnector";
-
-const transports = Object.fromEntries(
-  activeChains.map((chain) => [chain.id, http()]),
-);
-
-// Switch miniapps
-const connectors = [injected(), beexoConnector()];
-
-export const webConfig = createConfig({
-  chains: activeChains,
-  transports,
-  connectors,
-  ssr: true,
-});
+// This file is being depopulated in favor of adapters/
+// Only keeping this if there are other shared exports needed, otherwise it can be empty or removed.
+// Currently empty as webConfig moved to adapters/web.ts
+export { };
