+++ tsconfig.json
@@ -1,7 +1,11 @@
 {
   "compilerOptions": {
     "target": "ES2020",
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -19,7 +23,9 @@
       },
     ],
     "paths": {
-      "@/*": ["./src/*"],
+      "@/*": [
+        "./src/*"
+      ],
     },
   },
   "include": [
@@ -29,5 +35,11 @@
     ".next/types/**/*.ts",
     ".next/dev/types/**/*.ts",
   ],
-  "exclude": ["node_modules", "scripts", "test", "hardhat.config.ts"],
+  "exclude": [
+    "node_modules",
+    "scripts",
+    "test",
+    "hardhat.config.ts",
+    "contracts"
+  ],
 }
